requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pytest>=7.4.0
httpx>=0.24.0
python-multipart>=0.0.6
redis>=4.5.4
fastapi-cache2>=0.2.1
slowapi>=0.1.8
aioredis>=2.0.1
certifi>=2024.2.2